[gemspec] Remove upper bounds on versions for all dependencies (#333)
Let's optimistically assume that all future versions will be compatible.
We can add restrictions or roll out a fix if that turns out not to be
the case.

The upside of this change is that we don't have to make changes to this
gem's gemspec to accommodate newly released versions of its
dependencies.